+++ concordium-sdk/src/main/java/com/concordium/sdk/Client.java
@@ -2,38 +2,37 @@ package com.concordium.sdk;
 
 import com.concordium.sdk.exceptions.*;
 import com.concordium.sdk.responses.AccountIndex;
-import com.concordium.sdk.responses.peerlist.Peer;
-import com.concordium.sdk.responses.peerStats.PeerStatistics;
-import com.concordium.sdk.responses.nodeinfo.NodeInfo;
-import com.concordium.sdk.responses.blocksatheight.BlocksAtHeight;
-import com.concordium.sdk.exceptions.AccountNotFoundException;
-import com.concordium.sdk.exceptions.BlockNotFoundException;
-import com.concordium.sdk.exceptions.TransactionNotFoundException;
-import com.concordium.sdk.exceptions.TransactionRejectionException;
 import com.concordium.sdk.responses.accountinfo.AccountInfo;
 import com.concordium.sdk.responses.blockinfo.BlockInfo;
+import com.concordium.sdk.responses.blocksatheight.BlocksAtHeight;
 import com.concordium.sdk.responses.blocksatheight.BlocksAtHeightRequest;
 import com.concordium.sdk.responses.blocksummary.BlockSummary;
 import com.concordium.sdk.responses.consensusstatus.ConsensusStatus;
 import com.concordium.sdk.responses.cryptographicparameters.CryptographicParameters;
 import com.concordium.sdk.responses.getancestors.BlockAncestors;
+import com.concordium.sdk.responses.nodeinfo.NodeInfo;
+import com.concordium.sdk.responses.peerStats.PeerStatistics;
+import com.concordium.sdk.responses.peerlist.Peer;
 import com.concordium.sdk.responses.transactionstatus.TransactionStatus;
-import com.concordium.sdk.transactions.*;
+import com.concordium.sdk.transactions.AccountAddress;
+import com.concordium.sdk.transactions.AccountNonce;
+import com.concordium.sdk.transactions.Hash;
+import com.concordium.sdk.transactions.Transaction;
+import com.concordium.sdk.types.UInt16;
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.ByteString;
-import org.semver4j.Semver;
+import com.google.protobuf.Int32Value;
 import concordium.ConcordiumP2PRpc;
 import concordium.P2PGrpc;
 import io.grpc.ManagedChannel;
 import lombok.val;
+import org.semver4j.Semver;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
 import java.time.Duration;
-import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 /**
  * The Client is responsible for sending requests to the node.
@@ -327,9 +326,34 @@ public final class Client {
     }
 
     /**
+     * Ask the node to join the specified network.
+     *
+     * @param networkId {@link UInt16} Network ID.
+     * @return true if network has been joined successfully. False otherwise.
+     */
+    public boolean joinNetwork(final UInt16 networkId) {
+        return server().joinNetwork(ConcordiumP2PRpc.NetworkChangeRequest.newBuilder()
+                .setNetworkId(Int32Value.newBuilder().setValue(networkId.getValue()).build())
+                .build()).getValue();
+    }
+
+    /**
+     * Ask the node to leave the specified network.
+     *
+     * @param networkId {@link UInt16} Network ID.
+     * @return true if network has been left successfully. False otherwise.
+     */
+    public boolean leaveNetwork(final UInt16 networkId) {
+        return server().leaveNetwork(ConcordiumP2PRpc.NetworkChangeRequest.newBuilder()
+                .setNetworkId(Int32Value.newBuilder().setValue(networkId.getValue()).build())
+                .build()).getValue();
+    }
+
+    /**
      * Gets Block Ancestor Blocks.
+     *
      * @param blockHash {@link Hash} of the block.
-     * @param num Total no of Ancestor blocks to get.
+     * @param num       Total no of Ancestor blocks to get.
      * @return {@link BlockAncestors}
      * @throws BlockNotFoundException When the returned response from Node is invalid or null.
      */
@@ -339,7 +363,7 @@ public final class Client {
                         .newBuilder()
                         .setBlockHash(blockHash.asHex())
                         .setAmount(num)
-                .build());
+                        .build());
 
         val res = BlockAncestors.fromJson(jsonResponse);
         if (Objects.isNull(res)) {